backend-apps/auth/requirements.txt
aiomysql==0.2.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.31.0
azure-mgmt-core==1.4.0
azure-mgmt-storage==21.2.1
azure-storage-blob==12.23.0
bcrypt==4.1.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isodate==0.6.1
asyncmy==0.2.9
Jinja2==3.1.4
makefun==1.15.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msrest==0.7.1
oauthlib==3.2.2
orjson==3.10.6
packaging==24.1
pwdlib==0.2.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0